fix MinGw/Windows related issue with too long PlaceIDs

+++ prg1/datastructures.hh
@@ -18,8 +18,8 @@
 #include <memory>
 
 // Types for IDs
-using PlaceID = long int;
-using AreaID = long int;
+using PlaceID = long long int;
+using AreaID = long long int;
 using Name = std::string;
 using WayID = std::string;
